Recombine tests and also rename files + fix with new test
