+++ gems/aws-sdk-core/spec/aws-sdk-core/config_file_factory_spec.rb
