+++ gems/aws-sdk-core/spec/aws-sdk-core/config_file_section_spec.rb
